Refactor: 예외 처리 로직 개선
- 예외 메시지 가공의 책임: ResponseHandler -> GlobalExceptionHandler
- Message 설명 필드: message -> description
- Message: @Getter & @AllArgsConstructor 추가
