+++ src/main/java/com/onmoim/server/common/response/Message.java
@@ -1,0 +1,14 @@
+package com.onmoim.server.common.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum Message {
+	SUCCESS("SUCCESS"),
+	BAD_REQUEST("BAD REQUEST"),
+	SYSTEM_ERROR("SYSTEM ERROR");
+
+	private final String description;
+}
