+++ src/main/java/com/onmoim/server/common/response/ResponseHandler.java
@@ -1,13 +1,6 @@
 package com.onmoim.server.common.response;
 
 import java.time.LocalDateTime;
-import java.util.List;
-import java.util.stream.Collectors;
-
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-
-import com.onmoim.server.common.exception.CustomException;
 
 import lombok.Builder;
 import lombok.Data;
@@ -26,54 +19,15 @@ public class ResponseHandler<T> {
 
 	public static <T> ResponseHandler<T> response(T data) {
 		return ResponseHandler.<T>builder()
-			.message(Message.SUCCESS.message)
+			.message(Message.SUCCESS.getDescription())
 			.data(data)
 			.build();
 	}
 
-	public static <T> ResponseHandler<T> errorResponse(final CustomException customException) {
+	public static <T> ResponseHandler<T> errorResponse(T data, String message) {
 		return ResponseHandler.<T>builder()
-			.message(customException.getErrorCode().name())
-			.data((T)customException.getMessage())
+			.message(message)
+			.data(data)
 			.build();
-	}
-
-	public static <T> ResponseHandler<T> errorResponse(final MethodArgumentNotValidException exception) {
-		List<FieldErrorResponse> errors = exception.getBindingResult().getFieldErrors()
-			.stream().map(FieldErrorResponse::of)
-			.collect(Collectors.toList());
-
-		return ResponseHandler.<T>builder()
-			.message(Message.BAD_REQUEST.message)
-			.data((T)errors)
-			.build();
-	}
-
-	public static <T> ResponseHandler<T> errorResponse() {
-		return ResponseHandler.<T>builder()
-			.message(Message.SYSTEM_ERROR.message)
-			.build();
-	}
-
-	private record FieldErrorResponse(String field, String message, Object rejectedValue) {
-		public static FieldErrorResponse of(final FieldError fieldError) {
-			return new FieldErrorResponse(
-				fieldError.getField(),
-				fieldError.getDefaultMessage(),
-				fieldError.getRejectedValue()
-			);
-		}
-	}
-
-	private enum Message {
-		SUCCESS("SUCCESS"),
-		BAD_REQUEST("BAD REQUEST"),
-		SYSTEM_ERROR("SYSTEM ERROR");
-
-		private final String message;
-
-		Message(String message) {
-			this.message = message;
-		}
 	}
 }
